some test clean up, split out list karma tests, reuse some basic stuff. fix some a11y issues in list view to add in no violations. cant do the same for dashboard rn because its catching false violations based on the grid item scrim - says theres a contrast violation but its fine.
